pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalshi-gas"
version = "0.1.0"
description = "End-to-end gas market analytics pipeline with ETL, modeling, evaluation, and reporting."
authors = [{name = "Mateo Bodon"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "requests>=2.31",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.6",
]

[project.scripts]
kalshi-gas = "kalshi_gas.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
